# BookReview_API/BookReview_API/Controllers/BooksController.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.EntityFrameworkCore;
using BookReview_API;
using BookReview_API.Data;

namespace BookReview_API.Controllers
{
    [Route("api/[controller]")]
    [ApiController]
    public class BooksController : ControllerBase
    {
        private readonly BookReview_APIContext _context;

        public BooksController(BookReview_APIContext context)
        {
            _context = context;
        }

        // GET: api/Books
        [HttpGet]
        public async Task<ActionResult<IEnumerable<Book>>> GetBook()
        {
            return await _context.Book.ToListAsync();
        }

        // GET: api/Books/5
        [HttpGet("{id:int}")]
        public async Task<ActionResult<Book>> GetBook(int id)
        {
            var book = await _context.Book.FindAsync(id);

            if (book == null)
            {
                return NotFound();
            }

            return book;
        }


        // GET: api/search/
        [HttpGet("search/{str:alpha}")]
        public async Task<ActionResult<Book>> GetSearchResults(string str)
        {
            var books = await _context.Book
        .Where(b => EF.Functions.Like(b.Title, $"%{str}%") || EF.Functions.Like(b.Author, $"%{str}%"))
        .ToListAsync();

            if (!books.Any())
            {
                return NotFound();
            }

            return Ok(books);
        }


        // GET: api/Books/author
        [HttpGet("author/{author:alpha}")]
        public async Task<ActionResult<IEnumerable<Book>>> GetAuthorsBooks(string author)
        {
            var books = await _context.Book.Where(r => r.Author == author).ToListAsync();

            if (!books.Any())
            {
                return NotFound();
            }

            return Ok(books);
        }

        // GET: api/Books/5/reviews
        [HttpGet("{id}/reviews")]
        public async Task<ActionResult<IEnumerable<Review>>> GetBookReviews(int id)
        {
            var reviews = await _context.Review.Where(r => r.BookId == id).ToListAsync();

            if (!reviews.Any())
            {
                return NotFound();
            }

            return Ok(reviews);
        }

        // PUT: api/Books/5
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
        [HttpPut("{id:int}")]
        public async Task<IActionResult> PutBook(int id, Book book)
        {
            if (id != book.BookId)
            {
                return BadRequest();
            }

            _context.Entry(book).State = EntityState.Modified;

            try
            {
                await _context.SaveChangesAsync();
            }
            catch (DbUpdateConcurrencyException)
            {
                if (!BookExists(id))
                {
                    return NotFound();
                }
                else
                {
                    throw;
                }
            }

            return NoContent();
        }

        // POST: api/Books
        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
        [HttpPost]
        public async Task<ActionResult<Book>> PostBook(Book book)
        {
            _context.Book.Add(book);
            await _context.SaveChangesAsync();

            return CreatedAtAction("GetBook", new { id = book.BookId }, book);
        }

        // DELETE: api/Books/5
        [HttpDelete("{id}")]
        public async Task<IActionResult> DeleteBook(int id)
        {
            var book = await _context.Book.FindAsync(id);
            if (book == null)
            {
                return NotFound();
            }

            _context.Book.Remove(book);
            await _context.SaveChangesAsync();

            return NoContent();
        }

        private bool BookExists(int id)
        {
            return _context.Book.Any(e => e.BookId == id);
        }
    }
}